Move HttpListenerSession to a separate file
